Move tests around to better match app layout
